htdocs/requirements.txt
coreapi==1.32.3
Django==1.10
django-crispy-forms==1.6.0
django-rest-swagger==2.0.5
djangorestframework==3.4.6
itypes==1.1.0
MySQL-python==1.2.5
openapi-codec==1.0.0
pytz==2016.6.1
requests==2.11.1
simplejson==3.8.2
uritemplate==0.6
